Fixed Manifestation test and hence visibility of `dumpShortenedRelations`
